requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==1.11.17
django-silk==3.0.4
djangorestframework==3.9.4
gprof2dot==2017.9.19
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
pycodestyle==2.5.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
